дз к 6 уроку ретрофит
