+++ app/src/main/java/com/astro/weathernow/interfaces/OpenWeather.java
@@ -1,0 +1,13 @@
+package com.astro.weathernow.interfaces;
+
+
+import com.astro.weathernow.model.WeatherRequest;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Query;
+
+public interface OpenWeather {
+    @GET("data/2.5/weather")
+Call<WeatherRequest> loadWeather(@Query("q") String cityCountry, @Query("appid") String keyApi);
+}
