+++ app/src/main/java/com/astro/weathernow/model/Clouds.java
