+++ app/src/main/java/com/astro/weathernow/model/Coord.java
