+++ app/src/main/java/com/astro/weathernow/model/Main.java
