+++ app/src/main/java/com/astro/weathernow/model/Sys.java
