+++ app/src/main/java/com/astro/weathernow/model/Weather.java
