+++ app/src/main/java/com/astro/weathernow/model/Wind.java
