+++ app/src/main/res/layout/activity_main.xml
@@ -80,11 +80,5 @@
         android:textAppearance="?android:attr/textAppearanceMedium"
         />
 
-    <ProgressBar
-        android:id="@+id/loader"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"/>
-
 
 </RelativeLayout>
